app/src/main/java/com/slam/outcomehealthvideos/utils/DateTimeUtils.java
package com.slam.outcomehealthvideos.utils;

/**
 * Created by slam on 09/09/2019.
 */
public class DateTimeUtils {

    public static String convertMillieToHMmSs(long millie) {
        long seconds = (millie / 1000);
        long second = seconds % 60;
        long minute = (seconds / 60) % 60;
        long hour = (seconds / (60 * 60)) % 24;

        String result = "";
        if (hour > 0) {
            return String.format("%02d:%02d:%02d", hour, minute, second);
        }
        else {
            return String.format("%02d:%02d" , minute, second);
        }
    }
}
